+++ examples/functions/decode/main.tf
@@ -1,4 +1,6 @@
 terraform {
+  required_version = ">=1.8"
+
   required_providers {
     toml = {
       source  = "registry.terraform.io/tobotimus/toml"