+++ main.go
@@ -39,7 +39,6 @@ func main() {
 	flag.Parse()
 
 	opts := providerserver.ServeOpts{
-		// TODO: Update this string with the published name of your provider.
 		Address: "registry.terraform.io/tobotimus/toml",
 		Debug:   debug,
 	}